chatapp/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.1
bcrypt==3.2.0
Flask-SocketIO==5.1.1
python-socketio==5.3.0
PyNaCl==1.4.0         # For libsodium encryption
gunicorn==20.1.0      # For production
eventlet==0.33.0      # For handling WebSockets in production

